corrijo el contraataque

+++ tadp-dbz/src/dbz/Criterio.scala
@@ -1,6 +1,6 @@
 package dbz
 
-class Criterio (funcion: (Guerrero,Guerrero)=>Int) {
+class Criterio (funcion: (Guerrero,Guerrero)=>Float) {
   def analizar(atacante: Guerrero, defensor: Guerrero) = {
     funcion(atacante,defensor)
   }
